+++ CHANGELOG.md
@@ -8,11 +8,37 @@ Please choose versions by [Semantic Versioning](http://semver.org/).
 * MINOR version when you add functionality in a backwards-compatible manner, and
 * PATCH version when you make backwards-compatible bug fixes.
 
-## v1.4.0
+## v1.6.0
 
 - remove vendor
 - go mod update
 
+## v1.5.2
+
+- add namespace to Alert.Identifier
+
+## v1.5.1
+
+- fix Alert.Identifier
+- add test for Alert
+
+## v1.5.0
+
+- add AlertEventHandler
+- add ResourceEventHandler
+
+## v1.4.1
+
+- enable applyconfig
+
+## v1.4.0
+
+- use new k8s code generator
+
+## v1.3.4
+
+- go mod update
+ 
 ## v1.3.3
 
 - update to k8s v0.31.0